updated table and  used enum
    pick 339dee5 feat: updated schema
