+++ .env
@@ -3,4 +3,4 @@ POSTGRES_HOST=127.0.0.1
 POSTGRES_PORT=5432
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=root
-POSTGRES_DATABASE=linkedin
+POSTGRES_DATABASE=employee
