+++ src/Employee-management/models/post.interface.ts
@@ -1,12 +1,12 @@
 export interface FeedPost {
     id? : number;
-    body? : string;
-    createdAt? : Date;
+    
+    //image?: ImageBitmap;
     EmployeeName?: string;
     EmployeeAddress?: string;
-    department?: string;
+   //
     contact?: number;
     salary?: number;
     
-
+    createdAt? : Date;
 }
