--- src/feed/models/post.entity.ts
@@ -1,31 +1,0 @@
-import{Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-
-@Entity('feed_post')
-export class FeedPostEntity{
-    @PrimaryGeneratedColumn()
-         id : number;
-
-    @Column({ default : '' })
-        body : string;
-        
-    @Column({type:'timestamp', default:() => 'CURRENT_TIMESTAMP'})
-        createdAt : Date;
-
-    @Column({ default : '' })
-    EmployeeName: string;
-
-    @Column({ default : '' })
-    EmployeeAddress: string;
-
-    
-    @Column({ default : '' })
-    department: string;
-
-    @Column({ type: 'int' , default: null})
-    contact: number;
-
-    @Column({ type: 'int' ,default: null })
-    salary: number;
-
-}
