+++ src/main.ts
@@ -1,15 +1,9 @@
-import {
-    NestFactory
-}
-from '@nestjs/core';
-import {
-    AppModule
-}
-from './app.module';
+import { NestFactory } from '@nestjs/core';
+import { AppModule } from './app.module';
 
 async function bootstrap() {
-    const app = await NestFactory.create(AppModule);
-    app.setGlobalPrefix('api');
-    await app.listen(3000);
+  const app = await NestFactory.create(AppModule);
+  app.setGlobalPrefix('emp');
+  await app.listen(3000);
 }
 bootstrap();
